+++ src/PostOrder/Types/CancelSummary.php
@@ -25,8 +25,8 @@ namespace DTS\eBaySDK\PostOrder\Types;
  * @property \DTS\eBaySDK\PostOrder\Types\Amount $requestRefundAmount
  * @property \DTS\eBaySDK\PostOrder\Types\DateTime $sellerResponseDueDate
  * @property \DTS\eBaySDK\PostOrder\Types\DateTime $shipmentDate
- * @property string $state
- * @property string $status
+ * @property string $cancelState
+ * @property string $cancelStatus
  */
 class CancelSummary extends \DTS\eBaySDK\Types\BaseType
 {
@@ -112,17 +112,17 @@ class CancelSummary extends \DTS\eBaySDK\Types\BaseType
             'attribute' => false,
             'elementName' => 'shipmentDate'
         ],
-        'state' => [
+        'cancelState' => [
             'type' => 'string',
             'repeatable' => false,
             'attribute' => false,
-            'elementName' => 'state'
+            'elementName' => 'cancelState'
         ],
-        'status' => [
+        'cancelStatus' => [
             'type' => 'string',
             'repeatable' => false,
             'attribute' => false,
-            'elementName' => 'status'
+            'elementName' => 'cancelStatus'
         ]
     ];